requirements.txt
flask
pinecone-client
transformers
torch